Clean code after Ephem refactoring
