+++ lib/astronoby.rb
@@ -10,7 +10,6 @@ require "astronoby/epoch"
 require "astronoby/instant"
 require "astronoby/vector"
 require "astronoby/velocity"
-require "astronoby/astronomical_models/ephemeride_lunaire_parisienne"
 require "astronoby/astronomical_models/moon_phases_periodic_terms"
 require "astronoby/bodies/solar_system_body"
 require "astronoby/bodies/moon"
@@ -28,20 +27,16 @@ require "astronoby/coordinates/equatorial"
 require "astronoby/coordinates/horizontal"
 require "astronoby/corrections/light_time_delay"
 require "astronoby/aberration"
-require "astronoby/aberration2"
 require "astronoby/deflection"
 require "astronoby/equinox_solstice"
 require "astronoby/errors"
 require "astronoby/events/moon_phases"
-require "astronoby/events/observation_events"
-require "astronoby/events/rise_transit_set_iteration"
 require "astronoby/events/twilight_events"
 require "astronoby/events/celestial_event_calculator"
 require "astronoby/geocentric_parallax"
 require "astronoby/mean_obliquity"
 require "astronoby/moon_phase"
 require "astronoby/nutation"
-require "astronoby/nutation2"
 require "astronoby/observer"
 require "astronoby/precession"
 require "astronoby/reference_frame"
@@ -53,7 +48,6 @@ require "astronoby/reference_frames/topocentric"
 require "astronoby/refraction"
 require "astronoby/time/greenwich_sidereal_time"
 require "astronoby/time/local_sidereal_time"
-require "astronoby/util/astrodynamics"
 require "astronoby/util/maths"
 require "astronoby/util/time"
 require "astronoby/util/trigonometry"
