+++ lib/astronoby/constants.rb
@@ -15,7 +15,6 @@ module Astronoby
 
     SECONDS_PER_HOUR = SECONDS_PER_MINUTE * MINUTES_PER_HOUR
     SECONDS_PER_DAY = SECONDS_PER_HOUR * HOURS_PER_DAY
-    SECONDS_PER_DEGREE = SECONDS_PER_MINUTE * MINUTES_PER_DEGREE
     RADIAN_PER_HOUR = Math::PI / 12.0
 
     PI_IN_DEGREES = 180.0
