+++ spec/astronoby/aberration_spec.rb
@@ -1,51 +1,47 @@
 # frozen_string_literal: true
 
 RSpec.describe Astronoby::Aberration do
-  describe "::for_ecliptic_coordinates" do
-    it "returns apparent ecliptic coordinates" do
-      true_coordinates = Astronoby::Coordinates::Ecliptic.new(
-        latitude: Astronoby::Angle.from_dms(20, 30, 40),
-        longitude: Astronoby::Angle.from_dms(30, 40, 50)
+  include TestEphemHelper
+
+  describe "#corrected_position" do
+    it "returns the corrected position for Mars on 2025-04-01" do
+      time = Time.utc(2025, 4, 1)
+      instant = Astronoby::Instant.from_time(time)
+      ephem = test_ephem
+      earth = Astronoby::Earth.new(instant: instant, ephem: ephem)
+      mars = Astronoby::Mars.new(instant: instant, ephem: ephem)
+      earth_geometric_velocity = earth.geometric.velocity
+      mars_astrometric_position = mars.astrometric.position
+      aberration = described_class.new(
+        astrometric_position: mars_astrometric_position,
+        observer_velocity: earth_geometric_velocity
       )
 
-      apparent_coordinates = described_class.for_ecliptic_coordinates(
-        coordinates: true_coordinates,
-        epoch: Astronoby::Epoch::DEFAULT_EPOCH
-      )
+      corrected_position = aberration.corrected_position
 
-      expect(apparent_coordinates).to be_a(Astronoby::Coordinates::Ecliptic)
-      expect(apparent_coordinates.latitude).not_to(
-        eq(true_coordinates.latitude)
-      )
-      expect(apparent_coordinates.longitude).not_to(
-        eq(true_coordinates.longitude)
-      )
+      expect(corrected_position.to_a.map(&:km).map(&:round))
+        .to eq([-67181971, 140334599, 69466965])
+      # Skyfield: -67181974 140334600 69466966
     end
 
-    # Source:
-    #  Title: Practical Astronomy with your Calculator or Spreadsheet
-    #  Authors: Peter Duffett-Smith and Jonathan Zwart
-    #  Edition: Cambridge University Press
-    #  Chapter: 36 - Aberration
-    it "computes accurate apparent coordinates" do
-      true_coordinates = Astronoby::Coordinates::Ecliptic.new(
-        latitude: Astronoby::Angle.from_dms(-1, 32, 56.4),
-        longitude: Astronoby::Angle.from_dms(352, 37, 10.1)
-      )
-      time = Time.utc(1988, 9, 8)
-      epoch = Astronoby::Epoch.from_time(time)
-
-      apparent_coordinates = described_class.for_ecliptic_coordinates(
-        coordinates: true_coordinates,
-        epoch: epoch
+    it "returns the corrected position for Neptune on 2025-08-01" do
+      time = Time.utc(2025, 8, 1)
+      instant = Astronoby::Instant.from_time(time)
+      ephem = test_ephem
+      earth = Astronoby::Earth.new(instant: instant, ephem: ephem)
+      neptune = Astronoby::Neptune.new(instant: instant, ephem: ephem)
+      earth_geometric_velocity = earth.geometric.velocity
+      neptune_astrometric_position = neptune.astrometric.position
+      aberration = described_class.new(
+        astrometric_position: neptune_astrometric_position,
+        observer_velocity: earth_geometric_velocity
       )
 
-      expect(apparent_coordinates.latitude.str(:dms)).to(
-        eq("-1° 32′ 56.3319″")
-      )
-      expect(apparent_coordinates.longitude.str(:dms)).to(
-        eq("+352° 37′ 30.4522″")
-      )
+      corrected_position = aberration.corrected_position
+
+      expect(corrected_position.to_a.map(&:km).map(&:round))
+        .to eq([4375222148, 154854265, -45253746])
+      # Skyfield: 4375222144 154854263 -45253747
     end
   end
 end
