+++ spec/astronoby/equinox_solstice_spec.rb
@@ -1,9 +1,13 @@
 # frozen_string_literal: true
 
 RSpec.describe Astronoby::EquinoxSolstice do
+  include TestEphemHelper
+
   describe ".new" do
     it "raises an error if the event is not supported" do
-      expect { described_class.new(2024, 4) }.to(
+      ephem = double
+
+      expect { described_class.new(2024, 4, ephem) }.to(
         raise_error(
           Astronoby::UnsupportedEventError,
           "Expected a format between 0, 1, 2, 3, got 4"
@@ -20,10 +24,11 @@ RSpec.describe Astronoby::EquinoxSolstice do
     #  Chapter: 6 - The Sun, p.143
     it "it returns the time for the 2004 March equinox" do
       year = 2004
+      ephem = larger_test_ephem
 
-      equinox = described_class.march_equinox(year)
+      equinox = described_class.march_equinox(year, ephem)
 
-      expect(equinox).to eq Time.utc(2004, 3, 20, 6, 47, 24)
+      expect(equinox).to eq Time.utc(2004, 3, 20, 6, 48, 38)
       # Time from Celestial Calculations: 2004-03-20T06:42:36
       # Time from Astronomical Algorithms: 2004-03-20T06:49:42
       # Time from IMCCE: 2004-03-20T06:48:38
@@ -31,10 +36,11 @@ RSpec.describe Astronoby::EquinoxSolstice do
 
     it "it returns the time for the 2024 March equinox" do
       year = 2024
+      ephem = larger_test_ephem
 
-      equinox = described_class.march_equinox(year)
+      equinox = described_class.march_equinox(year, ephem)
 
-      expect(equinox).to eq Time.utc(2024, 3, 20, 3, 5, 8)
+      expect(equinox).to eq Time.utc(2024, 3, 20, 3, 6, 24)
       # Time from IMCCE: 2024-03-20T03:06:24
     end
   end
@@ -47,10 +53,11 @@ RSpec.describe Astronoby::EquinoxSolstice do
     #  Chapter: 6 - The Sun, p.143
     it "it returns the time for the 2004 June solstice" do
       year = 2004
+      ephem = larger_test_ephem
 
-      equinox = described_class.june_solstice(year)
+      equinox = described_class.june_solstice(year, ephem)
 
-      expect(equinox).to eq Time.utc(2004, 6, 21, 0, 54, 29)
+      expect(equinox).to eq Time.utc(2004, 6, 21, 0, 56, 55)
       # Time from Celestial Calculations: 2004-03-21T00:49:41
       # Time from Astronomical Algorithms: 2004-03-21T00:57:57
       # Time from IMCCE: 2004-06-21T00:56:52
@@ -58,10 +65,11 @@ RSpec.describe Astronoby::EquinoxSolstice do
 
     it "it returns the time for the 2024 June solstice" do
       year = 2024
+      ephem = larger_test_ephem
 
-      equinox = described_class.june_solstice(year)
+      equinox = described_class.june_solstice(year, ephem)
 
-      expect(equinox).to eq Time.utc(2024, 6, 20, 20, 50, 18)
+      expect(equinox).to eq Time.utc(2024, 6, 20, 20, 51, 2)
       # Time from IMCCE: 2024-06-20T20:51:00
     end
   end
@@ -74,10 +82,11 @@ RSpec.describe Astronoby::EquinoxSolstice do
     #  Chapter: 6 - The Sun, p.143
     it "it returns the time for the 2004 September equinox" do
       year = 2004
+      ephem = larger_test_ephem
 
-      equinox = described_class.september_equinox(year)
+      equinox = described_class.september_equinox(year, ephem)
 
-      expect(equinox).to eq Time.utc(2004, 9, 22, 16, 31, 14)
+      expect(equinox).to eq Time.utc(2004, 9, 22, 16, 29, 51)
       # Time from Celestial Calculations: 2004-09-22T16:30:54
       # Time from Astronomical Algorithms: 2004-09-22T16:27:20
       # Time from IMCCE: 2004-09-22T16:29:50
@@ -85,10 +94,11 @@ RSpec.describe Astronoby::EquinoxSolstice do
 
     it "it returns the time for the 2024 September equinox" do
       year = 2024
+      ephem = larger_test_ephem
 
-      equinox = described_class.september_equinox(year)
+      equinox = described_class.september_equinox(year, ephem)
 
-      expect(equinox).to eq Time.utc(2024, 9, 22, 12, 38, 20)
+      expect(equinox).to eq Time.utc(2024, 9, 22, 12, 43, 41)
       # Time from IMCCE: 2024-09-22T12:43:40
     end
   end
@@ -101,10 +111,11 @@ RSpec.describe Astronoby::EquinoxSolstice do
     #  Chapter: 6 - The Sun, p.143
     it "it returns the time for the 2004 December solstice" do
       year = 2004
+      ephem = larger_test_ephem
 
-      equinox = described_class.december_solstice(year)
+      equinox = described_class.december_solstice(year, ephem)
 
-      expect(equinox).to eq Time.utc(2004, 12, 21, 12, 47, 42)
+      expect(equinox).to eq Time.utc(2004, 12, 21, 12, 41, 34)
       # Time from Celestial Calculations: 2004-12-21T12:44:22
       # Time from Astronomical Algorithms: 2004-12-21T12:42:40
       # Time from IMCCE: 2004-12-21T12:41:36
@@ -112,10 +123,11 @@ RSpec.describe Astronoby::EquinoxSolstice do
 
     it "it returns the time for the 2024 December solstice" do
       year = 2024
+      ephem = larger_test_ephem
 
-      equinox = described_class.december_solstice(year)
+      equinox = described_class.december_solstice(year, ephem)
 
-      expect(equinox).to eq Time.utc(2024, 12, 21, 9, 15, 19)
+      expect(equinox).to eq Time.utc(2024, 12, 21, 9, 20, 33)
       # Time from IMCCE: 2024-12-21T09:20:34
     end
   end
