+++ spec/astronoby/events/twilight_events_spec.rb
@@ -1,38 +1,39 @@
 # frozen_string_literal: true
 
 RSpec.describe Astronoby::Events::TwilightEvents do
+  include TestEphemHelper
+
   describe "#morning_civil_twilight_time" do
     it "returns a time" do
-      sun = Astronoby::Sun.new(time: Time.new)
+      time = Time.new
+      instant = Astronoby::Instant.from_time(time)
+      ephem = test_ephem
       observer = Astronoby::Observer.new(
         latitude: Astronoby::Angle.zero,
         longitude: Astronoby::Angle.zero
       )
-      twilight_events = described_class.new(observer: observer, sun: sun)
+      twilight_events = described_class.new(
+        observer: observer,
+        instant: instant,
+        ephem: ephem
+      )
 
       expect(twilight_events.morning_civil_twilight_time).to be_a(Time)
     end
 
     it "returns when the morning civil twilight starts" do
-      sun = Astronoby::Sun.new(time: Time.utc(1979, 9, 7))
-      observer = Astronoby::Observer.new(
-        latitude: Astronoby::Angle.from_degrees(52),
-        longitude: Astronoby::Angle.zero
-      )
-      twilight_events = described_class.new(observer: observer, sun: sun)
-
-      expect(twilight_events.morning_civil_twilight_time)
-        .to eq Time.utc(1979, 9, 7, 4, 47, 13)
-      # Time from IMCCE: 04:46
-    end
-
-    it "returns when the morning civil twilight starts" do
-      sun = Astronoby::Sun.new(time: Time.utc(2024, 3, 14))
+      time = Time.utc(2024, 3, 14)
+      instant = Astronoby::Instant.from_time(time)
+      ephem = larger_test_ephem
       observer = Astronoby::Observer.new(
         latitude: Astronoby::Angle.from_dms(-33, 52, 4),
         longitude: Astronoby::Angle.from_dms(151, 12, 26)
       )
-      twilight_events = described_class.new(observer: observer, sun: sun)
+      twilight_events = described_class.new(
+        observer: observer,
+        instant: instant,
+        ephem: ephem
+      )
 
       expect(twilight_events.morning_civil_twilight_time)
         .to eq Time.utc(2024, 3, 14, 19, 28, 0)
@@ -41,12 +42,18 @@ RSpec.describe Astronoby::Events::TwilightEvents do
 
     context "when the civil twilight doesn't start" do
       it "returns nil" do
-        sun = Astronoby::Sun.new(time: Time.utc(2024, 6, 20))
+        time = Time.utc(2024, 6, 20)
+        instant = Astronoby::Instant.from_time(time)
+        ephem = larger_test_ephem
         observer = Astronoby::Observer.new(
           latitude: Astronoby::Angle.from_degrees(65),
           longitude: Astronoby::Angle.zero
         )
-        twilight_events = described_class.new(observer: observer, sun: sun)
+        twilight_events = described_class.new(
+          observer: observer,
+          instant: instant,
+          ephem: ephem
+        )
 
         expect(twilight_events.morning_civil_twilight_time).to be_nil
       end
@@ -55,36 +62,35 @@ RSpec.describe Astronoby::Events::TwilightEvents do
 
   describe "#evening_civil_twilight_time" do
     it "returns a time" do
-      sun = Astronoby::Sun.new(time: Time.new)
+      time = Time.new
+      instant = Astronoby::Instant.from_time(time)
+      ephem = larger_test_ephem
       observer = Astronoby::Observer.new(
         latitude: Astronoby::Angle.zero,
         longitude: Astronoby::Angle.zero
       )
-      twilight_events = described_class.new(observer: observer, sun: sun)
+      twilight_events = described_class.new(
+        observer: observer,
+        instant: instant,
+        ephem: ephem
+      )
 
       expect(twilight_events.evening_civil_twilight_time).to be_a(Time)
     end
 
     it "returns when the evening civil twilight ends" do
-      sun = Astronoby::Sun.new(time: Time.utc(1979, 9, 7))
-      observer = Astronoby::Observer.new(
-        latitude: Astronoby::Angle.from_degrees(52),
-        longitude: Astronoby::Angle.zero
-      )
-      twilight_events = described_class.new(observer: observer, sun: sun)
-
-      expect(twilight_events.evening_civil_twilight_time)
-        .to eq Time.utc(1979, 9, 7, 19, 9, 8)
-      # Time from IMCCE: 19:10
-    end
-
-    it "returns when the evening civil twilight ends" do
-      sun = Astronoby::Sun.new(time: Time.utc(2024, 3, 14))
+      time = Time.utc(2024, 3, 14)
+      instant = Astronoby::Instant.from_time(time)
+      ephem = larger_test_ephem
       observer = Astronoby::Observer.new(
         latitude: Astronoby::Angle.from_dms(-33, 52, 4),
         longitude: Astronoby::Angle.from_dms(151, 12, 26)
       )
-      twilight_events = described_class.new(observer: observer, sun: sun)
+      twilight_events = described_class.new(
+        observer: observer,
+        instant: instant,
+        ephem: ephem
+      )
 
       expect(twilight_events.evening_civil_twilight_time)
         .to eq Time.utc(2024, 3, 14, 8, 39, 45)
@@ -93,12 +99,18 @@ RSpec.describe Astronoby::Events::TwilightEvents do
 
     context "when the civil twilight doesn't start" do
       it "returns nil" do
-        sun = Astronoby::Sun.new(time: Time.utc(2024, 6, 20))
+        time = Time.utc(2024, 6, 20)
+        instant = Astronoby::Instant.from_time(time)
+        ephem = larger_test_ephem
         observer = Astronoby::Observer.new(
           latitude: Astronoby::Angle.from_degrees(65),
           longitude: Astronoby::Angle.zero
         )
-        twilight_events = described_class.new(observer: observer, sun: sun)
+        twilight_events = described_class.new(
+          observer: observer,
+          instant: instant,
+          ephem: ephem
+        )
 
         expect(twilight_events.evening_civil_twilight_time).to be_nil
       end
@@ -107,36 +119,35 @@ RSpec.describe Astronoby::Events::TwilightEvents do
 
   describe "#morning_nautical_twilight_time" do
     it "returns a time" do
-      sun = Astronoby::Sun.new(time: Time.new)
+      time = Time.new
+      instant = Astronoby::Instant.from_time(time)
+      ephem = larger_test_ephem
       observer = Astronoby::Observer.new(
         latitude: Astronoby::Angle.zero,
         longitude: Astronoby::Angle.zero
       )
-      twilight_events = described_class.new(observer: observer, sun: sun)
+      twilight_events = described_class.new(
+        observer: observer,
+        instant: instant,
+        ephem: ephem
+      )
 
       expect(twilight_events.morning_nautical_twilight_time).to be_a(Time)
     end
 
     it "returns when the morning nautical twilight starts" do
-      sun = Astronoby::Sun.new(time: Time.utc(1979, 9, 7))
-      observer = Astronoby::Observer.new(
-        latitude: Astronoby::Angle.from_degrees(52),
-        longitude: Astronoby::Angle.zero
-      )
-      twilight_events = described_class.new(observer: observer, sun: sun)
-
-      expect(twilight_events.morning_nautical_twilight_time)
-        .to eq Time.utc(1979, 9, 7, 4, 5, 7)
-      # Time from IMCCE: 04:03
-    end
-
-    it "returns when the morning nautical twilight starts" do
-      sun = Astronoby::Sun.new(time: Time.utc(2024, 3, 14))
+      time = Time.utc(2024, 3, 14)
+      instant = Astronoby::Instant.from_time(time)
+      ephem = larger_test_ephem
       observer = Astronoby::Observer.new(
         latitude: Astronoby::Angle.from_dms(-33, 52, 4),
         longitude: Astronoby::Angle.from_dms(151, 12, 26)
       )
-      twilight_events = described_class.new(observer: observer, sun: sun)
+      twilight_events = described_class.new(
+        observer: observer,
+        instant: instant,
+        ephem: ephem
+      )
 
       expect(twilight_events.morning_nautical_twilight_time)
         .to eq Time.utc(2024, 3, 14, 18, 58, 49)
@@ -145,12 +156,18 @@ RSpec.describe Astronoby::Events::TwilightEvents do
 
     context "when the nautical twilight doesn't start" do
       it "returns nil" do
-        sun = Astronoby::Sun.new(time: Time.utc(2024, 6, 20))
+        time = Time.utc(2024, 6, 20)
+        instant = Astronoby::Instant.from_time(time)
+        ephem = larger_test_ephem
         observer = Astronoby::Observer.new(
           latitude: Astronoby::Angle.from_degrees(55),
           longitude: Astronoby::Angle.zero
         )
-        twilight_events = described_class.new(observer: observer, sun: sun)
+        twilight_events = described_class.new(
+          observer: observer,
+          instant: instant,
+          ephem: ephem
+        )
 
         expect(twilight_events.morning_nautical_twilight_time).to be_nil
       end
@@ -159,50 +176,55 @@ RSpec.describe Astronoby::Events::TwilightEvents do
 
   describe "#evening_nautical_twilight_time" do
     it "returns a time" do
-      sun = Astronoby::Sun.new(time: Time.new)
+      time = Time.new
+      instant = Astronoby::Instant.from_time(time)
+      ephem = larger_test_ephem
       observer = Astronoby::Observer.new(
         latitude: Astronoby::Angle.zero,
         longitude: Astronoby::Angle.zero
       )
-      twilight_events = described_class.new(observer: observer, sun: sun)
+      twilight_events = described_class.new(
+        observer: observer,
+        instant: instant,
+        ephem: ephem
+      )
 
       expect(twilight_events.evening_nautical_twilight_time).to be_a(Time)
     end
 
     it "returns when the evening nautical twilight ends" do
-      sun = Astronoby::Sun.new(time: Time.utc(1979, 9, 7))
-      observer = Astronoby::Observer.new(
-        latitude: Astronoby::Angle.from_degrees(52),
-        longitude: Astronoby::Angle.zero
-      )
-      twilight_events = described_class.new(observer: observer, sun: sun)
-
-      expect(twilight_events.evening_nautical_twilight_time)
-        .to eq Time.utc(1979, 9, 7, 19, 51, 14)
-      # Time from IMCCE: 19:52
-    end
-
-    it "returns when the evening nautical twilight ends" do
-      sun = Astronoby::Sun.new(time: Time.utc(2024, 3, 14))
+      time = Time.utc(2024, 3, 14)
+      instant = Astronoby::Instant.from_time(time)
+      ephem = larger_test_ephem
       observer = Astronoby::Observer.new(
         latitude: Astronoby::Angle.from_dms(-33, 52, 4),
         longitude: Astronoby::Angle.from_dms(151, 12, 26)
       )
-      twilight_events = described_class.new(observer: observer, sun: sun)
+      twilight_events = described_class.new(
+        observer: observer,
+        instant: instant,
+        ephem: ephem
+      )
 
       expect(twilight_events.evening_nautical_twilight_time)
-        .to eq Time.utc(2024, 3, 14, 9, 8, 56)
+        .to eq Time.utc(2024, 3, 14, 9, 8, 55)
       # Time from IMCCE: 09:08:37
     end
 
     context "when the nautical twilight doesn't start" do
       it "returns nil" do
-        sun = Astronoby::Sun.new(time: Time.utc(2024, 6, 20))
+        time = Time.utc(2024, 6, 20)
+        instant = Astronoby::Instant.from_time(time)
+        ephem = larger_test_ephem
         observer = Astronoby::Observer.new(
           latitude: Astronoby::Angle.from_degrees(55),
           longitude: Astronoby::Angle.zero
         )
-        twilight_events = described_class.new(observer: observer, sun: sun)
+        twilight_events = described_class.new(
+          observer: observer,
+          instant: instant,
+          ephem: ephem
+        )
 
         expect(twilight_events.evening_nautical_twilight_time).to be_nil
       end
@@ -211,42 +233,35 @@ RSpec.describe Astronoby::Events::TwilightEvents do
 
   describe "#morning_astronomical_twilight_time" do
     it "returns a time" do
-      sun = Astronoby::Sun.new(time: Time.new)
+      time = Time.new
+      instant = Astronoby::Instant.from_time(time)
+      ephem = larger_test_ephem
       observer = Astronoby::Observer.new(
         latitude: Astronoby::Angle.zero,
         longitude: Astronoby::Angle.zero
       )
-      twilight_events = described_class.new(observer: observer, sun: sun)
+      twilight_events = described_class.new(
+        observer: observer,
+        instant: instant,
+        ephem: ephem
+      )
 
       expect(twilight_events.morning_astronomical_twilight_time).to be_a(Time)
     end
 
-    # Source:
-    #  Title: Practical Astronomy with your Calculator or Spreadsheet
-    #  Authors: Peter Duffett-Smith and Jonathan Zwart
-    #  Edition: Cambridge University Press
-    #  Chapter: 50 - Twilight, p.114
     it "returns when the morning astronomical twilight starts" do
-      sun = Astronoby::Sun.new(time: Time.utc(1979, 9, 7))
-      observer = Astronoby::Observer.new(
-        latitude: Astronoby::Angle.from_degrees(52),
-        longitude: Astronoby::Angle.zero
-      )
-      twilight_events = described_class.new(observer: observer, sun: sun)
-
-      expect(twilight_events.morning_astronomical_twilight_time)
-        .to eq Time.utc(1979, 9, 7, 3, 19, 20)
-      # Time from Practical Astronomy: 03:12
-      # Time from IMCCE: 03:17
-    end
-
-    it "returns when the morning astronomical twilight starts" do
-      sun = Astronoby::Sun.new(time: Time.utc(2024, 3, 14))
+      time = Time.utc(2024, 3, 14)
+      instant = Astronoby::Instant.from_time(time)
+      ephem = larger_test_ephem
       observer = Astronoby::Observer.new(
         latitude: Astronoby::Angle.from_dms(-33, 52, 4),
         longitude: Astronoby::Angle.from_dms(151, 12, 26)
       )
-      twilight_events = described_class.new(observer: observer, sun: sun)
+      twilight_events = described_class.new(
+        observer: observer,
+        instant: instant,
+        ephem: ephem
+      )
 
       expect(twilight_events.morning_astronomical_twilight_time)
         .to eq Time.utc(2024, 3, 14, 18, 29, 12)
@@ -255,12 +270,18 @@ RSpec.describe Astronoby::Events::TwilightEvents do
 
     context "when the astronomical twilight doesn't start" do
       it "returns nil" do
-        sun = Astronoby::Sun.new(time: Time.utc(2024, 6, 20))
+        time = Time.utc(2024, 6, 20)
+        instant = Astronoby::Instant.from_time(time)
+        ephem = larger_test_ephem
         observer = Astronoby::Observer.new(
           latitude: Astronoby::Angle.from_degrees(49),
           longitude: Astronoby::Angle.zero
         )
-        twilight_events = described_class.new(observer: observer, sun: sun)
+        twilight_events = described_class.new(
+          observer: observer,
+          instant: instant,
+          ephem: ephem
+        )
 
         expect(twilight_events.morning_astronomical_twilight_time).to be_nil
       end
@@ -269,56 +290,55 @@ RSpec.describe Astronoby::Events::TwilightEvents do
 
   describe "#evening_astronomical_twilight_time" do
     it "returns a time" do
-      sun = Astronoby::Sun.new(time: Time.new)
+      time = Time.new
+      instant = Astronoby::Instant.from_time(time)
+      ephem = larger_test_ephem
       observer = Astronoby::Observer.new(
         latitude: Astronoby::Angle.zero,
         longitude: Astronoby::Angle.zero
       )
-      twilight_events = described_class.new(observer: observer, sun: sun)
+      twilight_events = described_class.new(
+        observer: observer,
+        instant: instant,
+        ephem: ephem
+      )
 
       expect(twilight_events.evening_astronomical_twilight_time).to be_a(Time)
     end
 
-    # Source:
-    #  Title: Practical Astronomy with your Calculator or Spreadsheet
-    #  Authors: Peter Duffett-Smith and Jonathan Zwart
-    #  Edition: Cambridge University Press
-    #  Chapter: 50 - Twilight, p.114
     it "returns when the evening astronomical twilight ends" do
-      sun = Astronoby::Sun.new(time: Time.utc(1979, 9, 7))
-      observer = Astronoby::Observer.new(
-        latitude: Astronoby::Angle.from_degrees(52),
-        longitude: Astronoby::Angle.zero
-      )
-      twilight_events = described_class.new(observer: observer, sun: sun)
-
-      expect(twilight_events.evening_astronomical_twilight_time)
-        .to eq Time.utc(1979, 9, 7, 20, 37, 1)
-      # Time from Practical Astronomy: 20:43
-      # Time from IMCCE: 20:37
-    end
-
-    it "returns when the evening astronomical twilight ends" do
-      sun = Astronoby::Sun.new(time: Time.utc(2024, 3, 14))
+      time = Time.utc(2024, 3, 14)
+      instant = Astronoby::Instant.from_time(time)
+      ephem = larger_test_ephem
       observer = Astronoby::Observer.new(
         latitude: Astronoby::Angle.from_dms(-33, 52, 4),
         longitude: Astronoby::Angle.from_dms(151, 12, 26)
       )
-      twilight_events = described_class.new(observer: observer, sun: sun)
+      twilight_events = described_class.new(
+        observer: observer,
+        instant: instant,
+        ephem: ephem
+      )
 
       expect(twilight_events.evening_astronomical_twilight_time)
-        .to eq Time.utc(2024, 3, 14, 9, 38, 33)
+        .to eq Time.utc(2024, 3, 14, 9, 38, 32)
       # Time from IMCCE: 09:38:17
     end
 
     context "when the astronomical twilight doesn't start" do
       it "returns nil" do
-        sun = Astronoby::Sun.new(time: Time.utc(2024, 6, 20))
+        time = Time.utc(2024, 6, 20)
+        instant = Astronoby::Instant.from_time(time)
+        ephem = larger_test_ephem
         observer = Astronoby::Observer.new(
           latitude: Astronoby::Angle.from_degrees(49),
           longitude: Astronoby::Angle.zero
         )
-        twilight_events = described_class.new(observer: observer, sun: sun)
+        twilight_events = described_class.new(
+          observer: observer,
+          instant: instant,
+          ephem: ephem
+        )
 
         expect(twilight_events.evening_astronomical_twilight_time).to be_nil
       end
@@ -327,12 +347,18 @@ RSpec.describe Astronoby::Events::TwilightEvents do
 
   describe "#time_for_zenith_angle" do
     it "returns a time" do
-      sun = Astronoby::Sun.new(time: Time.new)
+      time = Time.new
+      instant = Astronoby::Instant.from_time(time)
+      ephem = larger_test_ephem
       observer = Astronoby::Observer.new(
         latitude: Astronoby::Angle.zero,
         longitude: Astronoby::Angle.zero
       )
-      twilight_events = described_class.new(observer: observer, sun: sun)
+      twilight_events = described_class.new(
+        observer: observer,
+        instant: instant,
+        ephem: ephem
+      )
       zenith_angle = Astronoby::Angle.from_degrees(90 + 17)
 
       time = twilight_events.time_for_zenith_angle(
@@ -345,12 +371,18 @@ RSpec.describe Astronoby::Events::TwilightEvents do
 
     context "when the period of time is incompatible" do
       it "raises an error" do
-        sun = Astronoby::Sun.new(time: Time.new)
+        time = Time.new
+        instant = Astronoby::Instant.from_time(time)
+        ephem = larger_test_ephem
         observer = Astronoby::Observer.new(
           latitude: Astronoby::Angle.zero,
           longitude: Astronoby::Angle.zero
         )
-        twilight_events = described_class.new(observer: observer, sun: sun)
+        twilight_events = described_class.new(
+          observer: observer,
+          instant: instant,
+          ephem: ephem
+        )
         zenith_angle = Astronoby::Angle.zero
 
         expect {
