+++ spec/astronoby/true_obliquity_spec.rb
@@ -18,7 +18,7 @@ RSpec.describe Astronoby::TrueObliquity do
         epoch = Astronoby::Epoch.from_time(Time.utc(1987, 4, 10, 0, 0, 0))
         obliquity = described_class.for_epoch(epoch)
 
-        expect(obliquity.str(:dms)).to eq "+23° 26′ 36.7989″"
+        expect(obliquity.str(:dms)).to eq "+23° 26′ 36.8137″"
       end
     end
   end
